+++ src/array.rs
@@ -9,13 +9,68 @@
 
 //! Support for strategies producing fixed-length arrays.
 //!
-//! There is no explicit type for array strategies; instead, simply make an
-//! array containing the desired strategy(ies) of the desired length.
+//! An array of strategies (but only length 1 to 32 for now) is itself a
+//! strategy which generates arrays of that size drawing elements from the
+//! corresponding input strategies.
+//!
+//! See also [`UniformArrayStrategy`](struct.UniformArrayStrategy.html) for
+//! easily making a strategy for an array drawn from one strategy.
 //!
 //! General implementations are available for sizes 1 through 32.
 
+use std::marker::PhantomData;
+
 use strategy::*;
 use test_runner::*;
+
+/// A `Strategy` which generates fixed-size arrays containing values drawn from
+/// an inner strategy.
+///
+/// `T` must be an array type of length 1 to 32 whose values are produced by
+/// strategy `S`. Instances of this type are normally created by the various
+/// `uniformXX` functions in this module.
+///
+/// This is mainly useful when the inner strategy is not `Copy`, precluding
+/// expressing the strategy as `[myStrategy; 32]`, for example.
+///
+/// ## Example
+///
+/// ```
+/// #[macro_use] extern crate proptest;
+/// use proptest::prelude::*;
+///
+/// proptest! {
+///   #[test]
+///   fn test_something(a in prop::array::uniform32(1u32..)) {
+///     let unexpected = [0u32;32];
+///     // `a` is also a [u32;32], so we can compare them directly
+///     assert_ne!(unexpected, a);
+///   }
+/// }
+/// # fn main() { }
+/// ```
+#[derive(Clone, Copy, Debug)]
+pub struct UniformArrayStrategy<S, T> {
+    strategy: S,
+    _marker: PhantomData<T>,
+}
+
+impl<S, T> UniformArrayStrategy<S, T> {
+    /// Directly create a `UniformArrayStrategy`.
+    ///
+    /// This is only intended for advanced use, since the only way to specify
+    /// the array size is with the turbofish operator and explicitly naming the
+    /// type of the values in the array and the strategy itself.
+    ///
+    /// Prefer the `uniformXX` functions at module-level unless something
+    /// precludes their use.
+    pub fn new(strategy: S) -> Self {
+        UniformArrayStrategy {
+            strategy,
+            _marker: PhantomData,
+        }
+    }
+}
 
 /// A `ValueTree` operating over a fixed-size array.
 #[derive(Clone, Copy, Debug)]
@@ -26,7 +81,24 @@ pub struct ArrayValueTree<T> {
 }
 
 macro_rules! small_array {
-    ($n:tt : $($ix:expr),*) => {
+    ($n:tt $uni:ident : $($ix:expr),*) => {
+        /// Create a strategy to generate fixed-length arrays.
+        ///
+        /// All values within the new strategy are generated using the given
+        /// strategy. The length of the array corresponds to the suffix of the
+        /// name of this function.
+        ///
+        /// See [`UniformArrayStrategy`](struct.UniformArrayStrategy.html) for
+        /// example usage.
+        pub fn $uni<S : Strategy>
+            (strategy: S) -> UniformArrayStrategy<S, [ValueFor<S>; $n]>
+        {
+            UniformArrayStrategy {
+                strategy,
+                _marker: PhantomData
+            }
+        }
+
         impl<S : Strategy> Strategy for [S;$n] {
             type Value = ArrayValueTree<[S::Value;$n]>;
 
@@ -34,6 +106,23 @@ macro_rules! small_array {
                          -> Result<Self::Value, String> {
                 Ok(ArrayValueTree {
                     tree: [$(self[$ix].new_value(runner)?,)*],
+                    shrinker: 0,
+                    last_shrinker: None,
+                })
+            }
+        }
+
+        impl<S : Strategy> Strategy
+        for UniformArrayStrategy<S, [ValueFor<S>; $n]> {
+            type Value = ArrayValueTree<[S::Value; $n]>;
+
+            fn new_value(&self, runner: &mut TestRunner)
+                         -> Result<Self::Value, String> {
+                Ok(ArrayValueTree {
+                    tree: [$({
+                        let _ = $ix;
+                        self.strategy.new_value(runner)?
+                    },)*],
                     shrinker: 0,
                     last_shrinker: None,
                 })
@@ -77,52 +166,84 @@ macro_rules! small_array {
     }
 }
 
-small_array!(1: 0);
-small_array!(2: 0, 1);
-small_array!(3: 0, 1, 2);
-small_array!(4: 0, 1, 2, 3);
-small_array!(5: 0, 1, 2, 3, 4);
-small_array!(6: 0, 1, 2, 3, 4, 5);
-small_array!(7: 0, 1, 2, 3, 4, 5, 6);
-small_array!(8: 0, 1, 2, 3, 4, 5, 6, 7);
-small_array!(9: 0, 1, 2, 3, 4, 5, 6, 7, 8);
-small_array!(10: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9);
-small_array!(11: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
-small_array!(12: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11);
-small_array!(13: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12);
-small_array!(14: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13);
-small_array!(15: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14);
-small_array!(16: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15);
-small_array!(17: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-small_array!(18: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17);
-small_array!(19: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18);
-small_array!(20: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19);
-small_array!(21: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20);
-small_array!(22: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21);
-small_array!(23: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22);
-small_array!(24: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23);
-small_array!(25: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24);
-small_array!(26: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25);
-small_array!(27: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25, 26);
-small_array!(28: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25, 26, 27);
-small_array!(29: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28);
-small_array!(30: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29);
-small_array!(31: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30);
-small_array!(32: 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
-                 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31);
+small_array!(1 uniform1:
+             0);
+small_array!(2 uniform2:
+             0, 1);
+small_array!(3 uniform3:
+             0, 1, 2);
+small_array!(4 uniform4:
+             0, 1, 2, 3);
+small_array!(5 uniform5:
+             0, 1, 2, 3, 4);
+small_array!(6 uniform6:
+             0, 1, 2, 3, 4, 5);
+small_array!(7 uniform7:
+             0, 1, 2, 3, 4, 5, 6);
+small_array!(8 uniform8:
+             0, 1, 2, 3, 4, 5, 6, 7);
+small_array!(9 uniform9:
+             0, 1, 2, 3, 4, 5, 6, 7, 8);
+small_array!(10 uniform10:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9);
+small_array!(11 uniform11:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+small_array!(12 uniform12:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11);
+small_array!(13 uniform13:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12);
+small_array!(14 uniform14:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13);
+small_array!(15 uniform15:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14);
+small_array!(16 uniform16:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15);
+small_array!(17 uniform17:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
+small_array!(18 uniform18:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17);
+small_array!(19 uniform19:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18);
+small_array!(20 uniform20:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19);
+small_array!(21 uniform21:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20);
+small_array!(22 uniform22:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21);
+small_array!(23 uniform23:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22);
+small_array!(24 uniform24:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23);
+small_array!(25 uniform25:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24);
+small_array!(26 uniform26:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25);
+small_array!(27 uniform27:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25, 26);
+small_array!(28 uniform28:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25, 26, 27);
+small_array!(29 uniform29:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28);
+small_array!(30 uniform30:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29);
+small_array!(31 uniform31:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30);
+small_array!(32 uniform32:
+             0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
+             18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31);
 
 #[cfg(test)]
 mod test {